Crate iota_genesis_common

Functions

execute_genesis_transaction

Takes as input a transaction of kind Genesis and executes it. This can be used to obtain the TransactionEffects, TransactionEvents and Vec<Object> (vector of created objects) that result from the execution. Some EpochData, ProtocolConfig and LimitsMetrics are needed as input to be passed to the transaction executor.

get_genesis_protocol_config

Gets a ProtocolConfig for genesis based on a ProtocolVersion.

prepare_and_execute_genesis_transaction

Prepares the data necessary and then invokes execute_genesis_transaction. The chain_start_timestamp_ms is used to construct a genesis EpochData. The protocol_version is used to create a ProtocolConfig for genesis. The genesis_transaction is the transaction to be executed.