iota_genesis_common

Function execute_genesis_transaction

pub fn execute_genesis_transaction(
    epoch_data: &EpochData,
    protocol_config: &ProtocolConfig,
    metrics: Arc<LimitsMetrics>,
    genesis_transaction: &Transaction,
) -> (TransactionEffects, TransactionEvents, Vec<Object>)

Takes as input a transaction of kind Genesis and executes it. This can be used to obtain the TransactionEffects, TransactionEvents and Vec<Object> (vector of created objects) that result from the execution. Some EpochData, ProtocolConfig and LimitsMetrics are needed as input to be passed to the transaction executor.