iota_genesis_common

Trait MigrationTxDataExt

pub trait MigrationTxDataExt {
    // Required methods
    fn get_objects(&self) -> impl Iterator<Item = Object> + '_;
    fn objects_by_tx_digest(
        &self,
        digest: TransactionDigest,
    ) -> Option<Vec<Object>>;
}

Recovers the objects created by the migration transactions of a MigrationTxData by re-executing them. Executing needs the Move VM, so these operations live here instead of next to the data type in iota-config.

Required Methods

fn get_objects(&self) -> impl Iterator<Item = Object> + '_

Executes all the migration transactions for this migration data and returns the objects created by these executions.

fn objects_by_tx_digest(&self, digest: TransactionDigest) -> Option<Vec<Object>>

Executes the migration transaction identified by digest and returns the vector of objects created by the execution.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".